Creo el UserTrait  con el metodo havePermission y lo agrego a user. Creo el gate y blindo el RoleController

+++ app/Http/Controllers/RoleController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use App\ArgusPermission\Models\Role;
 use App\ArgusPermission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -18,6 +19,7 @@ class RoleController extends Controller
 
     public function index()
     {
+        Gate::authorize('haveaccess','role.index');
         $roles = Role::orderBy('id', 'Desc')->paginate(2);
         return view('role.index', compact('roles'));
     }
@@ -25,6 +27,7 @@ class RoleController extends Controller
 
     public function create()
     {
+        Gate::authorize('haveaccess','role.create');
         $permissions = Permission::get();
         return view('role.create', compact('permissions'));
     }
@@ -32,6 +35,7 @@ class RoleController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('haveaccess','role.create');
         //validacion
         $request->validate([
             'name' => 'required|max:50|unique:roles,name',
@@ -43,9 +47,9 @@ class RoleController extends Controller
         //Create role
         $role =Role::create($request->all());
     
-        if($request->get('permission')){
+        //if($request->get('permission')){
             $role->permissions()->sync($request->get('permission'));
-        }
+        //}
 
         return redirect()->route('role.index')->with('status_success', 'Role saved successfully');
     
@@ -55,6 +59,7 @@ class RoleController extends Controller
 
     public function show(Role $role)
     {
+        $this->authorize('haveaccess','role.show');
         $permission_role=[];
         foreach($role->permissions as $permission){
             $permission_role[]=$permission->id;
@@ -68,7 +73,7 @@ class RoleController extends Controller
 
     public function edit(Role $role)
     {
-
+        $this->authorize('haveaccess','role.edit');
         $permission_role=[];
         foreach($role->permissions as $permission){
             $permission_role[]=$permission->id;
@@ -83,6 +88,7 @@ class RoleController extends Controller
 
     public function update(Request $request, Role $role)
     {
+        $this->authorize('haveaccess','role.edit');
         //validacion
         $request->validate([
             'name' => 'required|max:50|unique:roles,name,'.$role->id,
@@ -94,9 +100,9 @@ class RoleController extends Controller
         //update role
         $role->update($request->all());
     
-        if($request->get('permission')){
+        //if($request->get('permission')){
             $role->permissions()->sync($request->get('permission'));
-        }
+        //}
 
         return redirect()->route('role.index')->with('status_success', 'Role updated successfully');
     
@@ -106,6 +112,7 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        $this->authorize('haveaccess','role.destroy');
         $role->delete();
         return redirect()->route('role.index')->with('status_success', 'Role successfully removed');
     }
